maintenance
No longer remove the events to autostart audio. This can/will be used to restart the audio on mobile devices when they leave the app and come back. This event resumes the audio.

Removed the comments to not clutter up the console.log of a mobile device. It can become laggy.