reformat
older devices *cough* iOS devices that are under 14.5 do not support this and prevent the game from loading. Reverting back to regular class members.
This makes this library more prone to error due to exposing the developer to private hash names they otherwise would not have access to. Thanks APPLE.